Introduce mktmpdir, use it.
Introduce a helper function to create a temporary directory, and use
it in PackageDetailsProvider to avoid the latter creating temporary
directories in the one where the script resides in (which can be
/usr/lib/node_modules when private-bower is installed globally)

Fixing https://github.com/Hacklone/private-bower/issues/196.

Max number of attempts at creating a temporary directory (with a
different name each time) is arbitratily set to 5.