Support new refreshDisabled setting for git and svn repository caches
The refreshDisabled setting allows for control over when the private-bower server actually calls out to the public registry for cached repositories. When refreshDisabled is true, we require that all updates to the repository cache must be performed manually. This can be used in situations where external access needs to be restricted and managed.